requirements.txt
aiogram>=3.0.0
apscheduler
pycryptodome
